+++ src/main/java/com/volonter/Bibilioteka/services/IzdavacService.java
@@ -3,6 +3,7 @@ package com.volonter.Bibilioteka.services;
 import com.volonter.Bibilioteka.entities.Autor;
 import com.volonter.Bibilioteka.entities.Izdavac;
 import com.volonter.Bibilioteka.repositories.IzdavacRepo;
+import com.volonter.Bibilioteka.repositories.KnjigaRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -13,6 +14,10 @@ import java.util.stream.StreamSupport;
 public class IzdavacService {
     @Autowired
     private IzdavacRepo izdavacRepo;
+
+    @Autowired
+    private KnjigaRepo knjigaRepo;
+
 
     public boolean kreirajIzadavaca(Izdavac izdavac){
         if(izdavac.getId() == null){
@@ -36,8 +41,12 @@ public class IzdavacService {
     }
 
 
-    //Opasno brisati izdavaca, moguce samo ukoliko nema knjige koje pripadaju datom izdavacu
-/*    public void izbrisiIzdavaca(Izdavac izdavac){
-        izdavacRepo.delete(izdavac);
-    }*/
+    public boolean izbrisiIzadavaca(Izdavac izdavac){
+        if(StreamSupport.stream(knjigaRepo.findKnjigasByIzdavac(izdavac).spliterator(),true).collect(Collectors.toList()).isEmpty()){
+            izdavacRepo.delete(izdavac);
+            return true;
+        }
+        else
+            return false;
+    }
 }
